replace Nir Soffer’s ASUnit with Lifepillar’s
Also, had to comment out the `logging_lib` property. Script Debugger
used to use the script’s path for `path to me` when compiling, but
not anymore. This makes it impossible to compile a script with a
property using `path to me`. This was actually the biggest reason for
the change in ASUnit versions. Because I had to make changes anyway.